Thread safe for object tenJ
* PROPOSAL tenJ *

Rationale:
`tenJ` is an object characterizing data associated with the tensor newton model and
is used in the recursive call to `nlls_iterate`. In order to make it thread safe,
all routines requiring this object will have to either use it through a new
argument to the subroutine or be passed along an existing derived type.

Since `nlls_iterate` has two possible calls 1) from `nlls_solve`, and, 2) from
`solve_newton_tensor`. In the first case eval_f is user supplied while in the
secont it is provided be RALFit and which makes use of `tenJ` object.
It is possible to add the derived type `tenJ_type` to the derived type
`nlls_workspace` as a new member.

Where available, `tenJ` (tenJ_type) will be referenced through `workspace` (nlls_workspace)
instance or is passed as a new subroutine argument (tenJ).
Also, in order to
accomodate this without altering the abstract interface of `eval_f_type`, the
newton_tensor version of param (`w%tparam`, of class tensor_params_type extended
from params_base_type) is enriched with a `tenJ_type` pointer to `tenJ` in such
a manner so as to make accesible `tenJ` from within `evaltensor_f` without
altering the abstract interface.