Add option to export COV matrix (var or JT*J)
This branch add support for an option to export the
Covariance Matrix, alternatively the variance vector or
the approximation to the Hessian matrix.

`options%save_covm = T/F` trigger the construction of the
Covariance Matrix.

Tests hace been added to check against certified values of the
problem Lanczos 3 and careful atention was given to test how the
Jacobian matrix is provided via option `options%fortran_jacobian=T/F`